Split serial driver in generic protocol and hardware driver part
This allows other means of split communication to be implemented by
simply implementing the driver specific methods used in the serial protocol.
